+++ designer/client/src/actions/reduxTypes.ts
@@ -3,6 +3,7 @@ import type { ThunkAction as TA, ThunkDispatch as TD } from "redux-thunk";
 
 import type { RootState } from "../reducers";
 import type { CloudDataActions } from "../reducers/cloudData";
+import type { SquashHistoryActions } from "../reducers/graph/historySquash";
 import type { ScenariosActions } from "../reducers/scenarios";
 import type { ActionTypes } from "./actionTypes";
 import type { CountsActions, NodeActions, NodeDetailsActions, PropertiesActions, ScenarioActions, SelectionActions } from "./nk";
@@ -27,6 +28,7 @@ type TypedAction =
     | ScenariosActions
     | SelectionActions
     | SettingsActions
+    | SquashHistoryActions
     | ToolbarActions
     | UiActions
     | UpdateScenarioActivitiesAction