+++ designer/client/src/components/graph/node-modal/editors/field/FieldSwitch.tsx
@@ -1,14 +1,17 @@
-import React, { useState, ReactNode, useMemo, useCallback } from "react";
-import { EditorType, ExpressionObj } from "../expression/types";
-import { Option } from "../../fragment-input-definition/TypeSelect";
-import { Box, Tabs, Tab, styled } from "@mui/material";
 import { css } from "@emotion/css";
-import { Editor, ParamType } from "../types";
-import { editorsParameters } from "../expression/editorsParameters";
-import { blendDarken, getBorderColor } from "../../../../../containers/theme/helpers";
-import { editors, isExtendedEditor } from "../expression/Editor";
+import { Box, Tabs, Tab, styled } from "@mui/material";
+import type { ReactNode } from "react";
+import React, { useState, useMemo, useCallback } from "react";
 import { useTranslation } from "react-i18next";
+
+import { blendDarken, getBorderColor } from "../../../../../containers/theme/helpers";
+import type { Option } from "../../fragment-input-definition/TypeSelect";
+import { editors, isExtendedEditor } from "../expression/Editor";
+import { editorsParameters } from "../expression/editorsParameters";
 import { InfoTooltip } from "../expression/InfoTooltip";
+import type { ExpressionObj } from "../expression/types";
+import { EditorType } from "../expression/types";
+import type { Editor, ParamType } from "../types";
 
 const StyledTab = styled(Tab)(({ theme }) => ({
     fontSize: "0.65rem",
@@ -87,7 +90,7 @@ export const FieldSwitch = ({ availableEditors, onValueChange, expressionObj, ch
         availableEditors.find((editor) => {
             const editorParameters = editorsParameters[editor.type];
 
-            return editorParameters.language === expressionObj.language && allowsSwitch(editor);
+            return editorParameters?.language === expressionObj?.language && allowsSwitch(editor);
         }) ?? availableEditors[0],
     );