+++ designer/client/src/components/toolbars/edit/buttons/UndoButton.tsx
@@ -1,27 +1,29 @@
 import React from "react";
 import { useTranslation } from "react-i18next";
 import { useSelector } from "react-redux";
+
 import Icon from "../../../../assets/img/toolbarButtons/undo.svg";
-import { getHistoryPast } from "../../../../reducers/selectors/graph";
+import { getHistoryCounts } from "../../../../reducers/selectors/getHistory";
 import { useSelectionActions } from "../../../graph/SelectionContextProvider";
-import { CapabilitiesToolbarButton } from "../../../toolbarComponents/CapabilitiesToolbarButton";
-import { ToolbarButtonProps } from "../../types";
+import type { ToolbarButtonProps } from "../../types";
+import { CounterToolbarButton } from "./CounterToolbarButton";
 
 function UndoButton(props: ToolbarButtonProps): JSX.Element {
     const { undo } = useSelectionActions();
-    const past = useSelector(getHistoryPast);
+    const [past] = useSelector(getHistoryCounts);
     const { t } = useTranslation();
     const { disabled, type } = props;
-    const available = !disabled && past.length > 0 && undo;
+    const available = !disabled && undo;
 
     return (
-        <CapabilitiesToolbarButton
+        <CounterToolbarButton
             editFrontend
             name={t("panels.actions.edit-undo.button", "undo")}
             disabled={!available}
             icon={<Icon />}
             onClick={available ? (e) => undo(e.nativeEvent) : null}
             type={type}
+            count={past}
         />
     );
 }