+++ designer/client/src/windowManager/useWindows.ts
@@ -1,13 +1,16 @@
 import type { WindowId, WindowType } from "@touk/window-manager";
 import { useWindowManager } from "@touk/window-manager";
 import { defaults } from "lodash";
-import { useCallback, useEffect, useMemo } from "react";
+import { useCallback, useEffect, useLayoutEffect, useMemo, useRef } from "react";
+import type { DefaultRootState } from "react-redux";
+import { useSelector } from "react-redux";
 
 import { useUserSettings } from "../common/userSettings";
 import { StickyNoteType } from "../components/graph/utils/stickyNotesUtils";
 import type { ConfirmDialogData } from "../components/modals/GenericConfirmDialog";
 import type { InfoDialogData } from "../components/modals/GenericInfoDialog";
 import type { Scenario } from "../components/Process/types";
+import type { RootState } from "../reducers";
 import type { NodeType } from "../types";
 import { WindowKind } from "./WindowKind";
 
@@ -34,6 +37,15 @@ const useRemoveFocusOnEscKey = (isWindowOpen: boolean) => {
         };
     }, [isWindowOpen]);
 };
+
+function useSelectorRef<S = DefaultRootState, T = unknown>(selector: (state: S) => T) {
+    const selected = useSelector(selector);
+    const result = useRef(selected);
+    useLayoutEffect(() => {
+        result.current = selected;
+    }, [selected]);
+    return result;
+}
 
 const DEFAULT_WINDOW_MARGIN = 30;
 export const DEFAULT_WINDOW_WIDTH = 820;
@@ -68,15 +80,21 @@ export function useWindows(parent?: WindowId) {
         [forceDisableModals, _open],
     );
 
+    const nodeWindowIdMap = useSelectorRef((state: RootState) => state.nodeWindowIdMap);
+
     const openNodeWindow = useCallback(
         (node: NodeType, scenario: Scenario, readonly?: boolean) => {
             if (node.type === StickyNoteType) return;
+            const windowId = nodeWindowIdMap.current[node.id] || node.id;
             return open({
-                id: node.id,
+                id: windowId,
                 title: node.id,
                 isResizable: true,
                 kind: readonly ? WindowKind.viewNode : WindowKind.editNode,
-                meta: { node, scenario },
+                meta: {
+                    node,
+                    scenario,
+                },
                 shouldCloseOnEsc: false,
                 layoutData: settings["node.showInputsAndOutputs"]
                     ? {
@@ -92,7 +110,7 @@ export function useWindows(parent?: WindowId) {
                       },
             });
         },
-        [open, settings],
+        [nodeWindowIdMap, open, settings],
     );
 
     const inform = useCallback(
@@ -110,7 +128,10 @@ export function useWindows(parent?: WindowId) {
             return open({
                 title: data.text,
                 kind: WindowKind.confirm,
-                meta: defaults(data, { confirmText: "Yes", denyText: "No" }),
+                meta: defaults(data, {
+                    confirmText: "Yes",
+                    denyText: "No",
+                }),
                 ...(data.width != null && { layoutData: { width: data.width } }),
             });
         },